functions.c: removed one_arg and two_arg multi.c: changed numberOfArgs to return an int instead of uint8_t operator.c: changed two_arg to use the function pointer instead 	    modulo had & instead of % operatorUtility.c: put in array initializer thing (i don't know what they're called), indenting too stack.c: free_stk, changed while loop to a for because decrement in an index scares me a little bit stack.h/types.h: moved the definitions of union fp and struct oper into types.h to make stack.h only for stack stuff

+++ multi.c
@@ -11,8 +11,8 @@
 #include "multi.h"
 
 
-uint8_t numberOfArgs(char **inp) {
-	uint8_t i = 0;
+int numberOfArgs(char **inp) {
+	int i = 0;
 	for(; inp[i]; ++i); //empty for
 	return i;
 }
@@ -294,7 +294,7 @@ struct matrix *solve(char **inp, struct vari *var, err_ret *error) {
 
 
 struct matrix *zeros(char **inp, struct vari *var, err_ret *error) {
-	uint8_t dim = numberOfArgs(inp);
+	int dim = numberOfArgs(inp);
 	uint16_t *size = NULL;
 
 	//only one inp, make a square matrix of that size
@@ -466,7 +466,7 @@ struct matrix *extractValue(char **inp, int var_ind, struct vari *var, err_ret *
 		return NULL;
 	}
 
-	uint8_t dim = numberOfArgs(inp);    
+	int dim = numberOfArgs(inp);    
 	struct matrix *out = NULL;
 
 	struct vari *tmp = cpy_var(var);
@@ -665,7 +665,7 @@ char *removeSpaces(char *inp) {
 err_ret printLine(char **inp, struct vari *var) {
 	err_ret error = 0;
 
-	uint8_t argNo = numberOfArgs(inp);
+	int argNo = numberOfArgs(inp);
 
 	struct vari *tmp = cpy_var(var);
 	if( !tmp )
